refactor: types test
