+++ package.json
@@ -13,7 +13,7 @@
     "lint": "standard | snazzy",
     "lint:fix": "standard --fix | snazzy",
     "pretest": "npm run clean",
-    "test": "npm run lint && npm run unit && npm run typescript && ts-node test/types/index.ts",
+    "test": "npm run lint && npm run unit && npm run typescript",
     "posttest": "npm run clean",
     "typescript": "tstyche --config tstyche.config.json",
     "prepare": "husky",
@@ -52,7 +52,6 @@
     "husky": "^9.1.7",
     "snazzy": "^9.0.0",
     "standard": "^17.1.2",
-    "ts-node": "^10.9.2",
     "tstyche": "^4.0.2",
     "typescript": "^5.8.3"
   },
