+++ tsconfig.tstyche.json
@@ -1,10 +1,9 @@
 {
-  "extends": "../tsconfig.json",
+  "extends": "fastify-tsconfig",
   "compilerOptions": {
+    "noUnusedLocals": false,
     "noEmit": true,
     "strict": true,
     "types": []
-  },
-  "include": ["./"],
-  "exclude": []
+  }
 }
